Show first two widgets on page.
In order to make this work, there are quite a few other additions as
well.

Updated the structure of the app by adding a widgets directory.
Created the TodoHeader and CreateTodo widgets
Added those to the TodosPage

The number of Todos wasn't displaying correctly; modified
active_todo_count_cubit and todo_list_cubit so the 3 hardcoded todos
show up in the counter.